Simple implementation of main methods.
Added implementation to methods countWords, countCheckSum, printHelp.
Current issues:
- supported delimiter is only space;
- the limit for counter is ULLONG_MAX;
- the checksum can be displayed as chars only.

+++ WordCounter.h
@@ -9,7 +9,7 @@ class WordCounter
 public:
 	WordCounter(void);
 	~WordCounter(void);
-	static long countWords(const std::string &filename, const std::string &value);
+	static unsigned long long countWords(const std::string &filename, const std::string &value);
 private:
 	static std::vector<std::string> split (const std::string & text);
 };
